refactor(users): change tutorId type from Integer to Long in UserResource.

+++ src/main/java/com/skillswap/platform/tutormatch/Users/Interfaces/rest/resources/UserResource.java
@@ -12,4 +12,4 @@ public record UserResource(Long id,
                            String gender,
                            int semester,
                            RoleType roleType,
-                           Integer tutorId) {}
+                           Long tutorId) {}
